+++ src/packages/piral-cli/src/apps/build-piral.ts
@@ -1,33 +1,27 @@
 import * as Bundler from 'parcel-bundler';
 import { join, dirname, basename } from 'path';
-import { extendConfig, setStandardEnvs, checkExists, findFile, getPiletsInfo } from './common';
+import { extendConfig, setStandardEnvs, retrievePiletsInfo } from './common';
 
 export interface BuildPiralOptions {
   entry?: string;
   target?: string;
+  publicUrl?: string;
 }
 
 export const buildPiralDefaults = {
   entry: './src/index.html',
   target: './dist/index.html',
+  publicUrl: '/',
 };
 
 export async function buildPiral(baseDir = process.cwd(), options: BuildPiralOptions = {}) {
-  const { entry = buildPiralDefaults.entry, target = buildPiralDefaults.target } = options;
+  const {
+    entry = buildPiralDefaults.entry,
+    target = buildPiralDefaults.target,
+    publicUrl = buildPiralDefaults.publicUrl,
+  } = options;
   const entryFiles = join(baseDir, entry);
-  const exists = await checkExists(entryFiles);
-
-  if (!exists) {
-    return console.error('The given entry pointing to "%s" does not exist.', entryFiles);
-  }
-
-  const packageJson = await findFile(entryFiles, 'package.json');
-
-  if (!packageJson) {
-    return console.error('Cannot find any package.json. You need a valid package.json for your Piral instance.');
-  }
-
-  const { externals } = getPiletsInfo(packageJson);
+  const { externals } = await retrievePiletsInfo(entryFiles);
 
   await setStandardEnvs({
     production: true,
@@ -44,6 +38,7 @@ export async function buildPiral(baseDir = process.cwd(), options: BuildPiralOpt
       minify: true,
       scopeHoist: false,
       contentHash: true,
+      publicUrl,
     }),
   );