+++ src/packages/piral-cli/src/apps/debug-piral.ts
@@ -1,20 +1,31 @@
 import { join, dirname } from 'path';
-import { runDebug } from './common';
+import { runDebug, retrievePiletsInfo } from './common';
 
 export interface DebugPiralOptions {
   entry?: string;
   port?: number;
+  publicUrl?: string;
 }
 
 export const debugPiralDefaults = {
   entry: './src/index.html',
   port: 1234,
+  publicUrl: '/',
 };
 
-export function debugPiral(baseDir = process.cwd(), options: DebugPiralOptions = {}) {
-  const { entry = debugPiralDefaults.entry, port = debugPiralDefaults.port } = options;
+export async function debugPiral(baseDir = process.cwd(), options: DebugPiralOptions = {}) {
+  const {
+    entry = debugPiralDefaults.entry,
+    port = debugPiralDefaults.port,
+    publicUrl = debugPiralDefaults.publicUrl,
+  } = options;
   const entryFiles = join(baseDir, entry);
+  const { externals } = await retrievePiletsInfo(entryFiles);
   return runDebug(port, entryFiles, {
-    target: dirname(entry),
+    publicUrl,
+    options: {
+      target: dirname(entry),
+      dependencies: externals,
+    },
   });
 }