+++ src/packages/piral-core/src/hooks/search.ts
@@ -1,5 +1,4 @@
 import { useEffect, useRef } from 'react';
-import { wrapElement } from 'react-arbiter';
 import { useDebounce } from './debounce';
 import { useGlobalState } from './globalState';
 import { useActions } from './actions';
@@ -15,35 +14,15 @@ import { Disposable } from '../types';
  * will be integrated as they arrive.
  */
 export function useSearch(): [string, (value: string) => void] {
-  const { setSearchInput, resetSearchResults, appendSearchResults } = useActions();
+  const { setSearchInput, triggerSearch } = useActions();
   const searchInput = useGlobalState(m => m.search.input);
-  const providers = useGlobalState(m => m.components.searchProviders);
-  const q = useDebounce(searchInput);
+  const query = useDebounce(searchInput);
   const cancel = useRef<Disposable>(undefined);
 
   useEffect(() => {
-    const providerKeys = Object.keys(providers);
-    const load = !!q && providerKeys.length > 0;
     cancel.current && cancel.current();
-    resetSearchResults(load);
-
-    if (load) {
-      let searchCount = providerKeys.length;
-      let active = true;
-      cancel.current = () => (active = false);
-      providerKeys.forEach(key =>
-        providers[key].search(q).then(
-          results => {
-            active && appendSearchResults(results.map(m => wrapElement(m)), --searchCount === 0);
-          },
-          ex => {
-            console.warn(ex);
-            active && --searchCount === 0 && appendSearchResults([], true);
-          },
-        ),
-      );
-    }
-  }, [q]);
+    cancel.current = triggerSearch(query, false);
+  }, [query]);
 
   return [searchInput, setSearchInput];
 }