+++ src/packages/piral/src/components/index.ts
@@ -1,3 +1,4 @@
+export * from './AppLayout';
 export * from './Dashboard';
 export * from './ErrorInfo';
 export * from './Menu';