+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Piral Changelog
 
+## 0.5.0 (to be determined)
+
+* Added more props to the `Layout`
+* Allow setting and extending the user
+* Fixed rendering of modals
+* Added `public-url` option in Piral CLI
+* Fixed `piral-cli` build for scoped shared dependencies
+* Added error code and output to Piral CLI
+* Added more options to the search provider registration
+* Provide immediate argument to search provider
+
 ## 0.4.0 (June 11, 2019)
 
 * Introduced `subscriptionUrl`